chore: move export const PairDetail = ({ pair, index }: PairDetailProps) => { to legacy
